# apps/dh/api-dh/source/DataHub.WebApi/Clients/ESettExchange/V1/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EsettExchangeWebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/esett/balance-responsible/{documentId}/storageDocument": {
      "get": {
        "tags": [
          "BalanceResponsible"
        ],
        "operationId": "storageDocument",
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/octet-stream": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/esett/{documentId}": {
      "get": {
        "tags": [
          "ESett"
        ],
        "operationId": "esett",
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeEventTrackingResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeEventTrackingResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeEventTrackingResult"
                }
              }
            }
          }
        }
      }
    },
    "/esett/{documentId}/dispatchDocument": {
      "get": {
        "tags": [
          "ESett"
        ],
        "operationId": "dispatchDocument",
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/octet-stream": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/esett/{documentId}/responseDocument": {
      "get": {
        "tags": [
          "ESett"
        ],
        "operationId": "responseDocument",
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/octet-stream": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/esett/download": {
      "post": {
        "tags": [
          "ESett"
        ],
        "operationId": "downloadPOST",
        "parameters": [
          {
            "name": "locale",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ExchangeEventDownloadFilter"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ExchangeEventDownloadFilter"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ExchangeEventDownloadFilter"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/csv": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalculationType": {
        "type": "string",
        "enum": [
          "BalanceFixing",
          "Aggregation"
        ]
      },
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "Received",
          "AwaitingDispatch",
          "AwaitingReply",
          "Accepted",
          "Rejected",
          "BizTalkAccepted"
        ]
      },
...
      "ExchangeEventDownloadFilter": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "filter": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ExchangeEventFilter"
              }
            ]
          },
          "sorting": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ExchangeEventSortPropertySorting"
              }
            ]
          }
        }
      },
      "ExchangeEventFilter": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "calculationType": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/CalculationType"
              }
            ]
          },
          "timeSeriesType": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeSeriesType"
              }
            ]
          },
          "gridAreaCode": {
            "type": "string",
            "nullable": true
          },
          "documentStatus": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/DocumentStatus"
              }
            ]
          },
          "periodFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "periodTo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdTo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "latestDispatchedFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "latestDispatchedTo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "documentId": {
            "type": "string",
            "nullable": true
          },
          "actorNumber": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ExchangeEventSortProperty": {
        "type": "string",
        "enum": [
          "CalculationType",
          "Created",
          "DocumentId",
          "DocumentStatus",
          "GridAreaCode",
          "TimeSeriesType",
          "LatestDispatched"
        ]
      },
      "ExchangeEventSortPropertySorting": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "sortProperty": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ExchangeEventSortProperty"
              }
            ]
          },
          "direction": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SortDirection"
              }
            ]
          }
        }
      },
...
      "ExchangeEventTrackingResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "documentId": {
            "type": "string"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "gridAreaCode": {
            "type": "string"
          },
          "calculationType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CalculationType"
              }
            ]
          },
          "timeSeriesType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeSeriesType"
              }
            ]
          },
          "periodFrom": {
            "type": "string",
            "format": "date-time"
          },
          "periodTo": {
            "type": "string",
            "format": "date-time"
          },
          "documentStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DocumentStatus"
              }
            ]
          },
          "lastDispatched": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "SortDirection": {
        "type": "string",
        "enum": [
          "Ascending",
          "Descending"
        ]
      },
      "TimeSeriesType": {
        "type": "string",
        "enum": [
          "MgaExchange",
          "Production",
          "Consumption"
        ]
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "scheme": "Bearer"
      }
    }
  },
  "security": [
    {
      "Bearer": [
        "Bearer"
      ]
    }
  ]
}